Extract common client services
