+++ src/renderer/services/binance/common.ts
@@ -9,10 +9,11 @@ import { map, mergeMap, shareReplay, distinctUntilChanged } from 'rxjs/operators
 import { eqOString } from '../../helpers/fp/eq'
 import { triggerStream } from '../../helpers/stateHelper'
 import { network$ } from '../app/service'
-import { GetExplorerTxUrl } from '../clients/types'
-import { ClientStateForViews } from '../types'
-import { getClient, getClientStateForViews } from '../utils'
-import { keystoreService } from '../wallet/common'
+import * as C from '../clients'
+import { ExplorerUrl$, GetExplorerTxUrl$ } from '../clients/types'
+import { ClientStateForViews } from '../clients/types'
+import { getClient, getClientStateForViews } from '../clients/utils'
+import { keystoreService } from '../wallet/keystore'
 import { getPhrase } from '../wallet/util'
 import { BinanceClientState, BinanceClientState$, Client$ } from './types'
 
@@ -76,20 +77,13 @@ const address$: Observable<O.Option<Address>> = client$.pipe(
 
 /**
  * Explorer url depending on selected network
- *
- * If a client is not available (e.g. by removing keystore), it returns `None`
- *
  */
-const explorerUrl$: Observable<O.Option<string>> = client$.pipe(
-  map(FP.pipe(O.map((client) => client.getExplorerUrl()))),
-  distinctUntilChanged(eqOString.equals),
-  shareReplay(1)
-)
+const explorerUrl$: ExplorerUrl$ = C.explorerUrl$(client$)
 
-const getExplorerTxUrl$: Observable<O.Option<GetExplorerTxUrl>> = client$.pipe(
-  map(FP.pipe(O.map((client) => client.getExplorerTxUrl))),
-  shareReplay(1)
-)
+/**
+ * Explorer url depending on selected network
+ */
+const getExplorerTxUrl$: GetExplorerTxUrl$ = C.getExplorerTxUrl$(client$)
 
 // `TriggerStream` to reload `Balances`
 const { stream$: reloadBalances$, trigger: reloadBalances } = triggerStream()
