+++ src/renderer/services/chain/client.ts
@@ -1,16 +1,17 @@
-import { XChainClient } from '@xchainjs/xchain-client'
 import { Chain } from '@xchainjs/xchain-util'
 import * as O from 'fp-ts/lib/Option'
-import { Observable } from 'rxjs'
 import * as Rx from 'rxjs'
+import * as RxOp from 'rxjs/operators'
 
+import * as BNC from '../binance'
 import * as BTC from '../bitcoin'
+import { XChainClient$ } from '../clients'
+import { Chain$ } from './types'
 
-export const clientByChain$ = (chain: Chain): Observable<O.Option<XChainClient>> => {
+export const clientByChain$ = (chain: Chain): XChainClient$ => {
   switch (chain) {
     case 'BNB':
-      // TODO (@Veado / @ThatStrangeGuy) Add BNB after https://github.com/thorchain/asgardex-electron/pull/583 has been merged
-      return Rx.of(O.none)
+      return BNC.client$
     case 'BTC':
       return BTC.client$
     case 'ETH':
@@ -21,3 +22,13 @@ export const clientByChain$ = (chain: Chain): Observable<O.Option<XChainClient>>
       return Rx.of(O.none)
   }
 }
+
+export const getClientByChain$: (chain$: Chain$) => XChainClient$ = (chain$) =>
+  chain$.pipe(
+    RxOp.switchMap(
+      O.fold(
+        () => Rx.EMPTY,
+        (chain) => clientByChain$(chain)
+      )
+    )
+  )
