+++ src/renderer/services/ethereum/types.ts
@@ -2,7 +2,7 @@ import { Client } from '@thorchain/asgardex-ethereum'
 import * as O from 'fp-ts/lib/Option'
 import * as Rx from 'rxjs'
 
-import { ClientState } from '../types'
+import { ClientState } from '../clients/types'
 
 export type Client$ = Rx.Observable<O.Option<Client>>
