+++ src/renderer/services/wallet/transaction.ts
@@ -1,5 +1,4 @@
 import * as RD from '@devexperts/remote-data-ts'
-import { Chain } from '@xchainjs/xchain-util'
 import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
 import * as Rx from 'rxjs'
@@ -8,62 +7,15 @@ import * as RxOp from 'rxjs/operators'
 import { observableState } from '../../helpers/stateHelper'
 import * as BNB from '../binance'
 import * as BTC from '../bitcoin'
-import { GetExplorerTxUrl } from '../clients/types'
-import { selectedAsset$ } from './common'
+import * as C from '../clients'
+import { ExplorerUrl$, GetExplorerTxUrl$ } from '../clients/types'
+import { client$, selectedAsset$ } from './common'
 import { INITIAL_LOAD_TXS_PROPS } from './const'
 import { ApiError, TxsPageLD, ErrorId, LoadTxsHandler, ResetTxsPageHandler, LoadTxsProps } from './types'
 
-const explorerUrlByChain$ = (chain: Chain): Rx.Observable<O.Option<string>> => {
-  switch (chain) {
-    case 'BNB':
-      return BNB.explorerUrl$
-    case 'BTC':
-      return BTC.explorerUrl$
-    case 'ETH':
-      // not implemented yet
-      return Rx.of(O.none)
-    case 'THOR':
-      // reload THOR balances - not available yet
-      return Rx.of(O.none)
-    default:
-      return Rx.of(O.none)
-  }
-}
+export const explorerUrl$: ExplorerUrl$ = C.explorerUrl$(client$)
 
-const explorerTxUrlByChain$ = (chain: Chain): Rx.Observable<O.Option<GetExplorerTxUrl>> => {
-  switch (chain) {
-    case 'BNB':
-      return BNB.getExplorerTxUrl$
-    case 'BTC':
-      return BTC.getExplorerTxUrl$
-    case 'ETH':
-      // not implemented yet
-      return Rx.of(O.none)
-    case 'THOR':
-      // reload THOR balances - not available yet
-      return Rx.of(O.none)
-    default:
-      return Rx.of(O.none)
-  }
-}
-
-export const explorerUrl$: Rx.Observable<O.Option<string>> = selectedAsset$.pipe(
-  RxOp.switchMap(
-    O.fold(
-      () => Rx.EMPTY,
-      ({ chain }) => explorerUrlByChain$(chain)
-    )
-  )
-)
-
-export const getExplorerTxUrl$: Rx.Observable<O.Option<GetExplorerTxUrl>> = selectedAsset$.pipe(
-  RxOp.switchMap(
-    O.fold(
-      () => Rx.EMPTY,
-      ({ chain }) => explorerTxUrlByChain$(chain)
-    )
-  )
-)
+export const getExplorerTxUrl$: GetExplorerTxUrl$ = C.getExplorerTxUrl$(client$)
 
 /**
  * State of `LoadTxsProps`, which triggers reload of txs history
